Switch the facility-level lat/lng to most recent
This commit changes the source of the latitude/longitude coordinates for
each facility from the RMPFac:tblFacility table to each facility's most
recent submission. According to a recent email from the EPA, RMP
considers the submission-level coordinates to be the authoritative ones,
while the RMPFac coordinates come from an older, static table.